3stars: add sanity check for rigid transform

+++ src/opencv/opencv.h
@@ -45,7 +45,7 @@ void cvTransfPoint(double *x, double *y, Homography Href, Homography Himg);
 
 void cvTransfH(Homography Href, Homography Himg, Homography *Hres);
 
-int cvCalculRigidTransform(s_star *star_array_img,
+double cvCalculRigidTransform(s_star *star_array_img,
 		struct s_star *star_array_ref, int n, Homography *Hom);
 
 #ifdef __cplusplus
